fix(skill): close authoring delivery gates
