+++ .buildchain/kfd/kfd-2/claims/agent-work-state-contract.json
@@ -52,7 +52,7 @@
       "pointer": {
         "kind": "file",
         "path": "framework/core/src/python/kungfu/agent/kfd3_api.registry.json",
-        "sha256": "71330a58acc7bda5e795b3bca9c0b478abe592daeca991b3062de4790d780ae7"
+        "sha256": "bd6b9a05e7df2e591fc857015f2828bd53b485ea43dcf1bba9cb864dda66a66a"
       },
       "description": "Agent collaboration-interface registration for the work-model query."
     }
@@ -79,7 +79,7 @@
       },
       {
         "path": "framework/core/src/python/kungfu/agent/kfd3_api.registry.json",
-        "sha256": "71330a58acc7bda5e795b3bca9c0b478abe592daeca991b3062de4790d780ae7"
+        "sha256": "bd6b9a05e7df2e591fc857015f2828bd53b485ea43dcf1bba9cb864dda66a66a"
       }
     ],
     "artifactSha256": [
