+++ .buildchain/kfd/kfd-2/release-claims.json
@@ -11,7 +11,7 @@
     "version": "4.0.0-alpha.1",
     "channel": "repository-default",
     "tag": "v4.0.0-alpha.1",
-    "sourceSha": "c051080a0cefa7f7bd70821ccb0e7e1286e6e0dd"
+    "sourceSha": "8a0ea60cf49b2bbd02b47d91854450fb50861b71"
   },
   "claims": [
     {
@@ -347,7 +347,7 @@
           "pointer": {
             "kind": "file",
             "path": "framework/core/src/python/kungfu/agent/kfd3_api.registry.json",
-            "sha256": "71330a58acc7bda5e795b3bca9c0b478abe592daeca991b3062de4790d780ae7"
+            "sha256": "bd6b9a05e7df2e591fc857015f2828bd53b485ea43dcf1bba9cb864dda66a66a"
           },
           "description": "Agent collaboration-interface registration for the work-model query."
         }
