+++ docs/adr/KF-ADR-019fee22-e71d-7da9-8a44-9403c21a5d62.md
@@ -5,20 +5,21 @@ adr_id: KF-ADR-019fee22-e71d-7da9-8a44-9403c21a5d62
 decision_status: accepted
 amends: [KF-ADR-019f86da-4f90-74c2-9cbb-24f1c34303bf]
 implementation_status: implemented
-implementation_prs: [https://github.com/kungfu-systems/kungfu/pull/2938, https://github.com/kungfu-systems/kungfu/pull/2948, https://github.com/kungfu-systems/kungfu/pull/2952, https://github.com/kungfu-systems/kungfu/pull/2958, https://github.com/kungfu-systems/kungfu/pull/2960, https://github.com/kungfu-systems/kungfu/pull/2983, https://github.com/kungfu-systems/kungfu/pull/2998, https://github.com/kungfu-systems/kungfu/pull/3007]
+implementation_prs: [https://github.com/kungfu-systems/kungfu/pull/2938, https://github.com/kungfu-systems/kungfu/pull/2948, https://github.com/kungfu-systems/kungfu/pull/2952, https://github.com/kungfu-systems/kungfu/pull/2958, https://github.com/kungfu-systems/kungfu/pull/2960, https://github.com/kungfu-systems/kungfu/pull/2983, https://github.com/kungfu-systems/kungfu/pull/2998, https://github.com/kungfu-systems/kungfu/pull/3007, https://github.com/kungfu-systems/kungfu/pull/3013]
 closure_pr: https://github.com/kungfu-systems/kungfu/pull/3007
-qualification_refs: [framework/skill/kungfu-skill.contract.json, framework/skill/schema/skill-definition-v2.schema.json, framework/skill/fixtures/contract-v2/cases.json, framework/skill/scripts/contract-v2.mjs, framework/skill/schema/skill-registry-state-v2.schema.json, framework/skill/schema/skill-lifecycle-plan-v2.schema.json, framework/skill/schema/skill-lifecycle-receipt-v2.schema.json, framework/skill/schema/skill-dependency-plan-v2.schema.json, framework/skill/schema/skill-dependency-invocation-receipt-v2.schema.json, framework/core/src/python/kungfu/agent/skill-decision.contract.json, framework/core/tests/python/test_skill_lifecycle.py, framework/core/tests/python/test_agent_skill_advisory.py]
+qualification_refs: [framework/skill/kungfu-skill.contract.json, framework/skill/schema/skill-definition-v2.schema.json, framework/skill/fixtures/contract-v2/cases.json, framework/skill/scripts/contract-v2.mjs, framework/skill/schema/skill-registry-state-v2.schema.json, framework/skill/schema/skill-lifecycle-plan-v2.schema.json, framework/skill/schema/skill-lifecycle-receipt-v2.schema.json, framework/skill/schema/skill-dependency-plan-v2.schema.json, framework/skill/schema/skill-dependency-invocation-receipt-v2.schema.json, framework/skill/schema/skill-authoring-spec-v1.schema.json, framework/skill/schema/skill-authoring-plan-v1.schema.json, framework/skill/schema/skill-authoring-receipt-v1.schema.json, framework/core/src/python/kungfu/agent/skill-decision.contract.json, framework/core/src/python/kungfu/skill/catalog.py, framework/core/tests/python/test_skill_lifecycle.py, framework/core/tests/python/test_skill_authoring.py, framework/core/tests/python/test_agent_skill_advisory.py]
 review_state: self-reviewed
 sensitivity: public
-last_reviewed: 2026-08-11
+last_reviewed: 2026-08-12
 ---
 
 # KF-ADR-019fee22-e71d-7da9-8a44-9403c21a5d62: Kungfu Skill v2 is a content-rooted Work-scoped orchestration contract
 
 - Status: accepted and implemented; the contract, canonical registry and
   lifecycle fold, thin surface bindings, native KFX/Profile admission, compact
-  Agent decision advisory, and deterministic qualification are retained, while
-  later authoring and multi-surface runtime waves remain outside this cut
+  Agent decision advisory, bounded workspace-local instruction-only authoring,
+  and deterministic qualification are retained, while later effectful authoring
+  and multi-surface runtime waves remain outside this cut
 - Date: 2026-08-11
 
 ## Amendment boundary
