+++ framework/core/src/python/kungfu/agent/kfd3_api.registry.json
@@ -1989,80 +1989,7 @@
       "projections": [
         "commands.json"
       ]
-    },
-    {
-      "id": "kungfu.skill.author.contract",
-      "surface": "cli",
-      "name": "kungfu skill author contract --json",
-      "purpose": "Discover the exact rooted workspace-local Skill authoring constraints, schemas, examples, actions, and non-claims.",
-      "maturity": "stable",
-      "visibility": "adjacent-agent",
-      "anchor": {
-        "kind": "catalog"
-      },
-      "projections": [
-        "commands.json",
-        "agent-pack-doc"
-      ]
-    },
-    {
-      "id": "kungfu.skill.author.catalog",
-      "surface": "cli",
-      "name": "kungfu skill author catalog --json",
-      "purpose": "Read the exact current Skill candidate catalog root required for mandatory deduplication.",
-      "maturity": "stable",
-      "visibility": "adjacent-agent",
-      "anchor": {
-        "kind": "catalog"
-      },
-      "projections": [
-        "commands.json",
-        "agent-pack-doc"
-      ]
-    },
-    {
-      "id": "kungfu.skill.author.inspect",
-      "surface": "cli",
-      "name": "kungfu skill author inspect --spec <spec.json> --json",
-      "purpose": "Compare one bounded Skill proposal with the current rooted catalog and fail closed on duplicates or ambiguity.",
-      "maturity": "stable",
-      "visibility": "adjacent-agent",
-      "anchor": {
-        "kind": "catalog"
-      },
-      "projections": [
-        "commands.json"
-      ]
-    },
-    {
-      "id": "kungfu.skill.author.scaffold",
-      "surface": "cli",
-      "name": "kungfu skill author scaffold --signals <signals.json> --spec <spec.json> --workspace <path> --target <relative-path> --json",
-      "purpose": "Preview or exact-root execute one deterministic new workspace-local instruction-only Skill draft without lifecycle mutation.",
-      "maturity": "stable",
-      "visibility": "adjacent-agent",
-      "anchor": {
-        "kind": "catalog"
-      },
-      "projections": [
-        "commands.json",
-        "agent-pack-doc"
-      ]
-    },
-    {
-      "id": "kungfu.skill.author.qualify",
-      "surface": "cli",
-      "name": "kungfu skill author qualify <draft-path> --json",
-      "purpose": "Qualify one exact draft as workspace-local instruction-only content without claiming installation, activation, or Work completion.",
-      "maturity": "stable",
-      "visibility": "adjacent-agent",
-      "anchor": {
-        "kind": "catalog"
-      },
-      "projections": [
-        "commands.json"
-      ]
-    },
+    }, {"id":"kungfu.skill.author.contract","surface":"cli","name":"kungfu skill author contract --json","purpose":"Discover the exact rooted workspace-local Skill authoring constraints, schemas, examples, actions, and non-claims.","maturity":"stable","visibility":"adjacent-agent","anchor":{"kind":"catalog"},"projections":["commands.json","agent-pack-doc"]}, {"id":"kungfu.skill.author.catalog","surface":"cli","name":"kungfu skill author catalog --json","purpose":"Read the exact current Skill candidate catalog root required for mandatory deduplication.","maturity":"stable","visibility":"adjacent-agent","anchor":{"kind":"catalog"},"projections":["commands.json","agent-pack-doc"]}, {"id":"kungfu.skill.author.inspect","surface":"cli","name":"kungfu skill author inspect --spec <spec.json> --json","purpose":"Compare one bounded Skill proposal with the current rooted catalog and fail closed on duplicates or ambiguity.","maturity":"stable","visibility":"adjacent-agent","anchor":{"kind":"catalog"},"projections":["commands.json"]}, {"id":"kungfu.skill.author.scaffold","surface":"cli","name":"kungfu skill author scaffold --signals <signals.json> --spec <spec.json> --workspace <path> --target <relative-path> --json","purpose":"Preview or exact-root execute one deterministic new workspace-local instruction-only Skill draft without lifecycle mutation.","maturity":"stable","visibility":"adjacent-agent","anchor":{"kind":"catalog"},"projections":["commands.json","agent-pack-doc"]}, {"id":"kungfu.skill.author.qualify","surface":"cli","name":"kungfu skill author qualify <draft-path> --json","purpose":"Qualify one exact draft as workspace-local instruction-only content without claiming installation, activation, or Work completion.","maturity":"stable","visibility":"adjacent-agent","anchor":{"kind":"catalog"},"projections":["commands.json"]},
     {
       "id": "kungfu.kfx.install",
       "surface": "cli",
