+++ framework/registry/contract.registry-envelope.json
@@ -17,7 +17,7 @@
       "id": "contract-registry",
       "source": "framework/contract/kungfu-contracts.registry.json",
       "artifact": "config/kungfu-contracts.registry.json",
-      "root": "sha256:95879bbe4a816ced66cee102014e6c4d80ee145d9465b6f648c17a6fcad425e5"
+      "root": "sha256:00ae9385e093fd6bbbe689e686cc8fda5e793b12146467f3deb0b1028ad37124"
     },
     {
       "id": "work-lifecycle-operation-matrix",
